Refactor benchmark: move some functions to onnx_exporter.py and benchmark_helper.py
