+++ onnxruntime/python/tools/transformers/benchmark_gpt2.py
@@ -17,7 +17,7 @@ import logging
 import torch
 import onnx
 from transformers import AutoConfig
-from gpt2_helper import Gpt2Helper, MyGPT2Model, MyGPT2LMHeadModel, MODEL_CLASSES, DEFAULT_TOLERANCE
+from gpt2_helper import Gpt2Helper, MODEL_CLASSES, DEFAULT_TOLERANCE
 from quantize_helper import QuantizeHelper
 from benchmark_helper import create_onnxruntime_session, setup_logger, prepare_environment, Precision
